+++ packages/frontend/src/components/complex-search/index.ts
@@ -1,0 +1,5 @@
+export { default as ComplexSearchBoundsSelect } from "./ComplexSearchBoundsSelect";
+export { default as ComplexSearchColorSelect } from "./ComplexSearchColorSelect";
+export { default as ComplexSearchDateSelect } from "./ComplexSearchDateSelect";
+export { default as ComplexSearchTab } from "./ComplexSearchTab";
+export { default as ComplexSearchUserSelect } from "./ComplexSearchUserSelect";
